A few fixes...
I added my MM config for all the stock wheels which will replace the
originals instead of adding new parts, and updated it with a MM patch
for the ERS wheels.  I also merged some more recent configuration for
the ERS wheels into the duplicate part config.  Mainly, the model
reference was not correct with how Alexustas packaged it up in his last
release.